Online Bill Pay With Credit or Debit Card/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
stripe==9.12.0
